# Conflicts:
#	.gradle/buildOutputCleanup/buildOutputCleanup.lock
#	.idea/caches/build_file_checksums.ser
#	.idea/workspace.xml
#	app/src/main/AndroidManifest.xml
#	app/src/main/java/pers/zhc/tools/floatingboard/FloatingBoardMainActivity.java
#	app/src/main/java/pers/zhc/tools/utils/DialogUtil.java
#	app/src/main/java/pers/zhc/tools/utils/PromptDialog.java
#	app/src/main/res/drawable/ic_fb.xml
#	app/src/main/res/drawable/ic_launcher_foreground.xml
#	app/src/main/res/layout/prompt_dialog_rl.xml

+++ app/src/main/java/pers/zhc/tools/epicycles/EpicyclesTest.java
@@ -53,7 +53,7 @@ public class EpicyclesTest extends BaseActivity {
         sb.setOnSeekBarChangeListener(new SeekBar.OnSeekBarChangeListener() {
             @Override
             public void onProgressChanged(SeekBar seekBar, int progress, boolean fromUser) {
-                epicyclesView.scale(((double) progress));
+                epicyclesView.scale(progress);
             }
 
             @Override
